requirements.txt
fastapi
uvicorn[standard]
gspread==5.12.4
oauth2client
typing_extensions
gunicorn
pytz
bcrypt
pyjwt
